Fix Teaspoon @ 1.1.5
Some weird things are happening with Teaspoon 1.2.0. It does not detect
DummyApp and as a result, cannot load its proper configuration. As a
result, Teaspoon tries to use PhantomJS instead of headless Chrome, and
fails miserably, as there's no PhantomJS (and should not be!) available
in the Docker image.

1.1.5 version somehowe works. Probably it's a bug introduced later.
Upstream solidus uses Teaspoon from master branch, which, possibly, is
another way to fix the issue. For now we'd stay on Teaspoon 1.1.5 until
further investigation.